🚚  fix : rename shoulderStretching file&function
